boqpro-python/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI / LLM
anthropic>=0.34.0

# Computer Vision (Phase 4+)
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0

# OCR (Phase 5+)
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0

# PDF Processing
PyMuPDF>=1.23.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.0.0
